+++ lib/dbf/format_profile.ex
@@ -9,7 +9,8 @@ defmodule DBF.FormatProfile do
     :header_layout,
     :field_descriptor_layout,
     :memo_family,
-    :record_layout
+    :record_layout,
+    :field_kinds
   ]
   defstruct [
     :version,
@@ -17,13 +18,15 @@ defmodule DBF.FormatProfile do
     :header_layout,
     :field_descriptor_layout,
     :memo_family,
-    :record_layout
+    :record_layout,
+    :field_kinds
   ]
 
   @type header_layout :: :foxbase_8 | :dbase_legacy_32
   @type field_descriptor_layout :: :foxbase_16 | :dbase_legacy_32
   @type memo_family :: :none | :dbt_iii | :dbt_iv
   @type record_layout :: :dbase_legacy
+  @type field_kind :: :character | :numeric | :float | :logical | :date | :text_memo
 
   @type t :: %__MODULE__{
           version: byte(),
@@ -31,8 +34,19 @@ defmodule DBF.FormatProfile do
           header_layout: header_layout(),
           field_descriptor_layout: field_descriptor_layout(),
           memo_family: memo_family(),
-          record_layout: record_layout()
+          record_layout: record_layout(),
+          field_kinds: %{binary() => field_kind()}
         }
+
+  @legacy_field_kinds %{
+    "C" => :character,
+    "N" => :numeric,
+    "F" => :float,
+    "L" => :logical,
+    "D" => :date
+  }
+
+  @memo_field_kinds Map.put(@legacy_field_kinds, "M", :text_memo)
 
   @profiles %{
     0x02 => %{
@@ -41,7 +55,8 @@ defmodule DBF.FormatProfile do
       header_layout: :foxbase_8,
       field_descriptor_layout: :foxbase_16,
       memo_family: :none,
-      record_layout: :dbase_legacy
+      record_layout: :dbase_legacy,
+      field_kinds: %{"C" => :character, "N" => :numeric}
     },
     0x03 => %{
       version: 0x03,
@@ -49,7 +64,8 @@ defmodule DBF.FormatProfile do
       header_layout: :dbase_legacy_32,
       field_descriptor_layout: :dbase_legacy_32,
       memo_family: :none,
-      record_layout: :dbase_legacy
+      record_layout: :dbase_legacy,
+      field_kinds: @legacy_field_kinds
     },
     0x83 => %{
       version: 0x83,
@@ -57,7 +73,8 @@ defmodule DBF.FormatProfile do
       header_layout: :dbase_legacy_32,
       field_descriptor_layout: :dbase_legacy_32,
       memo_family: :dbt_iii,
-      record_layout: :dbase_legacy
+      record_layout: :dbase_legacy,
+      field_kinds: @memo_field_kinds
     },
     0x8B => %{
       version: 0x8B,
@@ -65,7 +82,8 @@ defmodule DBF.FormatProfile do
       header_layout: :dbase_legacy_32,
       field_descriptor_layout: :dbase_legacy_32,
       memo_family: :dbt_iv,
-      record_layout: :dbase_legacy
+      record_layout: :dbase_legacy,
+      field_kinds: @memo_field_kinds
     }
   }